pyproject.toml
[project]
name = "nonebot_plugin_llob_master"
version = "1.0.0"
description = "Support for managing your LLOB on Windows"
authors = [
    {name = "KanbeReina", email = "kano.2525@qq.com"},
]
dependencies = [
    "nonebot2>=2.3.3",
    "nonebot-plugin-apscheduler>=0.5.0",
    "pywin32>=306",
    "psutil>=6.0.0",
    "httpx>=0.27.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
